make the build reproducible
This change makes the binaries easy to reproduce. This means that
it is possible to produce the same binary given the same version of
Go and the same git commit.

This change enables the verification of released binaries.

The following changes had to be made in order to do this:
- the build time had to be dropped
- the buildid had to be set via the ldflags
- trimpath had to be added to strip the GOPATH prefix from the source
file paths, thus not including a path specific to the host where the
binary was built
- the buildinfo output was modified to add the Go version & to remove
the build time

These changes help produce binaries which have the exact same hash for
the same version of Go.