Use SHA256 to digest the seccomp profile
The MD5 hash algorithm panics under GODEBUG=fips140=only. Calculate the
digest of the seccomp profile using SHA256 instead. The message
containing the digest is informational only so changing the hash
algorithm should not introduce any compatibility breaks. And as far as I
can tell the DockerOpt.msg field is a dead store so this change should
have no observable impact on behaviour aside from not panicking when run
in FIPS140-only mode.

Signed-off-by: Cory Snider <csnider@mirantis.com>